Proof of concept SIR generation by adjusting the LLVM codegen.
Previously I'd tried implementing SIR generation as a new codegen from
scratch. What I learned from this is:

 - There's a lot of boilerplate in the codegen API.

 - To get it right, you'd end up duplicating a lot of LLVM backend code,
   which would need to be kept in sync over time. This code often
   contains low-level details, which at this point in time I don't
   understand. The code is also hard to debug if you get it wrong.

 - Pretty much the whole codegen has to be implemented and working all
   at once. You cannot really incrementally implement parts.

I spent quite some time trying that approach. I generated about 2000
lines of backend code. It's slow going and relentless.

Due to frustration with all of the above, I took pause to see if I could
find a better way.

This change shows an idea I had at the end of last week. The general
idea is to -- rather than implement a whole new backend -- add calls to
SIR generation from within the existing LLVM backend. This means that:

 - We can build our SIR IR incrementally over time.

 - We don't have to understand all of the low-level details right off
   the bat.

 - There will be no code duplication and when the LLVM backend changes,
   instead of our code going out of sync we will get merge conflicts
   (in a good way).

 - The code should be smaller.

This all sounds good, but to make it work, there's one small detail. The
LLVM codegen is opaque-pointer-centric and such pointers are often used
as identifiers in LLVM API calls. For example you "add a new block to the
function behind the following opaque Value pointer". To generate SIR for
such a scenario we have to keep a mapping of LLVM pointers to the
corresponding SIR. For example so that we know which of the SIR
functions to add a block to. This will raise memory consumption a bit,
but I hope it isn't all that bad.

I'm raising a draft PR for comments before I proceed.